tests/wpt/web-platform-tests/trusted-types/block-string-assignment-to-Location-replace.tentative.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>

  <meta http-equiv="Content-Security-Policy" content="require-trusted-types">
</head>
<body>
<script>
  // TrustedURL replacements do not throw.
  test(t => {
    let p = createURL_policy(window, 1);
    let url = p.createURL(location.href + "#xxx");
    location.replace(url);
    assert_equals("" + url, location.href, "location href");
  }, "location.replace via policy (successful URL transformation).");

  // String replacements throw.
  test(t => {
    assert_throws(new TypeError(), _ => {
      location.replace("A string");
    });
  }, "`location.replace = string` throws");

  // Null replacement throws.
  test(t => {
    assert_throws(new TypeError(), _ => {
      location.replace(null);
    });
  }, "`location.replace = null` throws");
</script>
